# lang.node.d/bearserver.d/div.d/bearsapi_swagger_copy.yaml
swagger: '2.0'
info:
  version: '1.0.0'
  title: Swagger Bearsapi (Simple)
  description: A sample API that uses a bears=mongostore as an example to demonstrate features in the swagger-2.0 specification
  termsOfService: http://helloreverb.com/terms/
  contact:
    name: wolfhesse
    email: foo@example.com
    url: http://wolfhesse.chickenkiller.com
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
host: bearsapi.wolfspool.chickenkiller.com
basePath: /api
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
  /bears:
    get:
      description: Returns all bears from the system that the user has access to
      operationId: findBears
      produces:
        - application/json
        - application/xml
        - text/xml
        - text/html
      parameters:
        - name: tags
          in: query
          description: tags to filter by
          required: false
          type: array
          items:
            type: string
          collectionFormat: csv
        - name: limit
          in: query
          description: maximum number of results to return
          required: false
          type: integer
          format: int32
      responses:
        '200':
          description: pet response
          schema:
            type: array
            items:
              $ref: '#/definitions/bear'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
    post:
      description: Creates a new bear in the store.  Duplicates are allowed
      operationId: addBear
      produces:
        - application/json
      parameters:
        - name: bear
          in: body
          description: Bear to add to the store
          required: true
          schema:
            $ref: '#/definitions/newBear'
      responses:
        '200':
          description: bear response
          schema:
            $ref: '#/definitions/bear'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
  /bears/{id}:
    get:
      description: Returns a user based on a single ID, if the user does not have access to the pet
      operationId: findPetById
      produces:
        - application/json
        - application/xml
        - text/xml
        - text/html
      parameters:
        - name: id
          in: path
          description: ID of pet to fetch
          required: true
          type: string
      responses:
        '200':
          description: bear response
          schema:
            $ref: '#/definitions/bear'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
    delete:
      description: deletes a single pet based on the ID supplied
      operationId: deletePet
      parameters:
        - name: id
          in: path
          description: ID of pet to delete
          required: true
          type: string
      responses:
        '204':
          description: pet deleted
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
definitions:
  bear:
    type: object
    required:
      - id
      - name
    properties:
      id:
        type: string
      name:
        type: string
  newBear:
    type: object
    required:
      - name
    properties:
      name:
        type: string
  errorModel:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string